+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- `WalletService.sign` now uses `postMessage` instead of query param to transfer data. Data can now be string or byte array. Will return the type that was passed in.
+
 ## 0.9.0
 
 ### Added